requirements.txt
pywebview==4.4.1
Pillow==10.1.0
